Refactor ImportSpecctraSES to be usable in API
Currently there is pcbnew.ImportSpecctraSES() but it needs a pcbnew
frame set which makes it unusable outside of the scripting window or a
plugin.

This refactors actual exporting logic into separate helper and adds an
overload to ImportSpecctraSES(board, filename) that doesn't need pcbnew
window, instead it will import into the given board.